+++ R/contribution.R
@@ -1,6 +1,7 @@
 #' Calculate contribution metrics of sites and species
 #' 
-#' Cz metrics
+#' This function calculates metrics that assess the contribution of a given
+#' species to its bioregion.
 #' 
 #' @param cluster_object a `bioregion.clusters` object or a `data.frame` or a 
 #' list of `data.frame` containing multiple partitions. At least two partitions
@@ -11,6 +12,10 @@
 #' @param comat a co-occurrence `matrix` with sites as rows and species as
 #' columns. 
 #' 
+#' @param bipartite_link `NULL` by default. Needed for `Cz` indices. A
+#' `data.frame` where each row represents the interaction between two nodes
+#'  and an optional third column indicating the weight of the interaction.  
+#' 
 #' @param indices a `character` specifying the contribution metric to compute.
 #' Available options are `contribution`.
 #' 
@@ -19,10 +24,30 @@
 #' Its formula is the following:
 #' \eqn{(n_ij - ((n_i n_j)/n))/(sqrt(((n - n_j)/(n-1)) (1-(n_j/n)) ((n_i n_j)/n)))}
 #' 
-#' with n the number of sites, n_i the number of sites in which
-#' species i is present, n_j the number of sites belonging to the
-#' bioregion j, n_ij the number of occurrences of species i in sites
-#' belonging to the bioregion j.
+#' with \eqn{n} the number of sites, \eqn{n_i} the number of sites in which
+#' species \eqn{i} is present, \eqn{n_j} the number of sites belonging to the
+#' bioregion \eqn{j}, \eqn{n_ij} the number of occurrences of species \eqn{i}
+#' in sites belonging to the bioregion \eqn{j}.
+#' 
+#' `Cz` metrics are derived from \insertRef{Guimera2005}{bioregion}.
+#' Their respective formula are:
+#' \eqn{C_i = 1 - \sum_{s=1}^{N_M}{{(\frac{k_is}{k_i}})^2}}
+#' 
+#' where \eqn{k_{is}} is the number of links of node (species or site) \eqn{i}
+#' to nodes in bioregion \eqn{s}, and \eqn{k_i} is the total degree of node
+#' \eqn{i}. The participation coefficient of a node is therefore close to 1 if
+#' its links are uniformly distributed among all the bioregions and 0 if all
+#' its links are within its own bioregion.
+#' 
+#' And:
+#' \eqn{z_i = \frac{k_i - \overline{k_{si}}}{\sigma_{k_{si}}}}
+#' 
+#' where \eqn{k_i} is the number of links of node (species or site) \eqn{i} to
+#' other nodes in its bioregion \eqn{s_i}, \eqn{\overline{k_{si}}} is the
+#' average of \eqn{k} over all the nodes in \eqn{s_i}, and
+#' \eqn{\sigma_{k_{si}}} is the standard deviation of \eqn{k} in \eqn{s_i}.
+#' The within-bioregion degree z-score measures how well-connected node \eqn{i}
+#' is to other nodes in the bioregion.
 #'
 #' @seealso [partition_metrics]
 #' @return 
@@ -58,17 +83,30 @@
 #' contribution(cluster_object = com, comat = comat,
 #' indices = "contribution")
 #' 
+#' # Cz indices
+#' net_bip <- mat_to_net(comat, weight = TRUE)
+#' clust_bip <- netclu_greedy(net_bip, bipartite = TRUE)
+#' contribution(cluster_object = clust_bip, comat = comat, 
+#' bipartite_link = net_bip, indices = "Cz")
+#' 
 #' @export
 
 contribution <- function(cluster_object,
                          comat,
-                         indices = c("contribution", "Cz")){
+                         indices = c("contribution", "Cz"),
+                         bipartite_link = NULL){
   # 1. Controls ---------------------------------------------------------------
   # input can be of format bioregion.clusters
   if (inherits(cluster_object, "bioregion.clusters")) {
     if (inherits(cluster_object$clusters, "data.frame")) {
       has.clusters <- TRUE
       clusters <- cluster_object$clusters
+      
+      if(ncol(clusters) > 2) {
+        stop("This function is designed to be applied on a single partition.",
+             "Your cluster_object has multiple partitions (select only one).")
+      }
+      
     } else {
       if (cluster_object$name == "hierarchical_clustering") {
         stop("No clusters have been generated for your hierarchical tree,
@@ -84,18 +122,31 @@ contribution <- function(cluster_object,
          on a site x species matrix.")
   }
   
-  if(ncol(clusters) > 2) {
-    stop("This function is designed to be applied on a single partition.",
-         "Your cluster_object has multiple partitions (select only one).")
-  }
-  
   controls(args = NULL, data = comat, type = "input_matrix")
   
-  controls(args = indices, data = NULL, type = "character")
-  if(!(indices %in% c("contribution", "Cz"))){
+  # controls(args = indices, data = NULL, type = "character")
+  if (!is.character(indices)) {
+    stop(paste0(deparse(substitute(indices)),
+                " must be a character or a vector of characters."),
+         call. = FALSE
+    )
+  }
+  
+  if(!isTRUE(unique(indices %in% c("contribution", "Cz")))){
     stop("Please choose algorithm among the followings values:
     contribution or Cz.", call. = FALSE)
   }
+  
+  if("Cz" %in% indices && is.null(bipartite_link)){
+    stop("bipartite_link is needed to compute Cz indices.")
+  }
+  
+  if("Cz" %in% indices && cluster_object$inputs$bipartite == FALSE){
+    stop("Cz metrics can only be computed for a bipartite partition (where
+         both sites and species are assigned to a bioregion.")
+  }
+  
+  # Add controls for bipartite_link
   
   contribution_df <- NULL
   
@@ -103,8 +154,128 @@ contribution <- function(cluster_object,
   ## 2.1. Cz ------------------------------------------------------------------
   # only for bipartite cases; not implemented yet
   if("Cz" %in% indices){
-    message("Cz indices not yet implemented.")
-    Cz_df <- data.frame()
+    # cluster_object <- clust2 # remove
+    # bipartite_link <- net_bip # remove
+    
+    # Needs two data frames as inputs:
+    # bipartite_link is a data.frame of the links between nodes containing four
+    # columns: site, species, bioregion_site and bioregion_species 
+    # 
+    # bipartite_df contains three columns: node, bioregion and cat which
+    # respectively stand for the name of the node, its bioregion and its
+    # bipartite category
+    
+    # Rename columns Node1 as Sites and Node2 as Species, and throw warning if
+    # this is the case
+    if("Node1" %in% colnames(bipartite_link)){
+      colnames(bipartite_link)[colnames(bipartite_link) == "Node1"] <- "Site"
+      warning("Column 'Node1' has been renamed 'Sites'. If this column does not
+            correspond to sites, rename it before running the function.")
+    }
+    if("Node2" %in% colnames(bipartite_link)){
+      colnames(bipartite_link)[colnames(bipartite_link) == "Node2"] <- "Species"
+      warning("Column 'Node2' has been renamed 'Species'. If this column does not
+            correspond to species, rename it before running the function.")
+    }
+    
+    bipartite_df <- cluster_object$clusters
+    # Add a column category (site or species) to bipartite_df
+    bipartite_df$cat <- attributes(bipartite_df)$node_type
+    colnames(bipartite_df) <- c("Node", "Bioregion", "Category")
+    
+    # Add bioregions of the sites to the bipartite data.frame
+    bipartite_link$Site <- as.character(bipartite_link$Site)
+    bipartite_df$Node <- as.character(bipartite_df$Node)
+    bipartite_link <- dplyr::left_join(bipartite_link,
+                                       bipartite_df[, c("Node", "Bioregion")],
+                                       by = c("Site" = "Node"))
+    colnames(bipartite_link)[colnames(bipartite_link) == "Bioregion"] <-
+      "Bioregion_site"
+    
+    # Add bioregions of the species to the bipartite data.frame
+    bipartite_link$Species <- as.character(bipartite_link$Species)
+    bipartite_df$Node <- as.character(bipartite_df$Node)
+    bipartite_link <- dplyr::left_join(bipartite_link,
+                                       bipartite_df[, c("Node", "Bioregion")],
+                                       by = c("Species" = "Node"))
+    colnames(bipartite_link)[colnames(bipartite_link) == "Bioregion"] <-
+      "Bioregion_species"
+    
+    
+    # Compute coefficient of participation C
+    C_site <- data.frame()
+    dat_com <- bipartite_df[which(bipartite_df$Category == "site"), ]
+    for(i in 1:nrow(dat_com)){
+      tmp <-
+        table(bipartite_link[which(bipartite_link$Site == dat_com[i, "Node"]),
+                             "Bioregion_species"])
+      C_site <- rbind(C_site,
+                      data.frame(Node = dat_com[i, "Node"],
+                                 C = 1 - sum((tmp/sum(tmp))^2),
+                                 Category = "site"))
+    }
+    
+    C_sp <- data.frame()
+    dat_sp <- bipartite_df[which(bipartite_df$Category == "species"), ]
+    for(i in 1:nrow(dat_sp)){
+      tmp <-
+        table(bipartite_link[which(bipartite_link$Species == dat_sp[i, "Node"]),
+                             "Bioregion_site"])
+      C_sp <- rbind(C_sp,
+                    data.frame(Node = dat_sp[i, "Node"],
+                               C = 1 - sum((tmp/sum(tmp))^2),
+                               Category = "species"))
+    }
+    C_dat <- rbind(C_site, C_sp)
+    
+    # Merge results with bipartite_df
+    bipartite_df <- dplyr::left_join(bipartite_df, C_dat,
+                                     by = c("Node",  "Category"))
+    
+    # Compute z
+    bipartite_df$n_link_bioregion <- NA
+    for(i in 1:nrow(bipartite_df)){
+      if(bipartite_df[i, "Category"] == "site"){
+        tmp <- bipartite_link[which(bipartite_link$Site == 
+                                      bipartite_df[i, "Node"]), ]
+        bipartite_df[i, "n_link_bioregion"] <-
+          nrow(tmp[which(tmp$Bioregion_site == tmp$Bioregion_species), ])
+      } else{
+        tmp <- bipartite_link[which(bipartite_link$Species ==
+                                      bipartite_df[i, "Node"]), ]
+        bipartite_df[i, "n_link_bioregion"] <-
+          nrow(tmp[which(tmp$Bioregion_site == tmp$Bioregion_species), ])
+      }
+    }
+    
+    # Average number of links within a bioregion
+    mean_link_bioregion <- tapply(bipartite_df$n_link_bioregion,
+                                  bipartite_df$Bioregion,
+                                  mean)
+    mean_link_bioregion <-
+      data.frame(Bioregion = names(mean_link_bioregion),
+                 mean_link_bioregion = as.numeric(mean_link_bioregion))
+    bipartite_df <- dplyr::left_join(bipartite_df, mean_link_bioregion,
+                                     by = "Bioregion")
+    
+    # Standard deviation of the number of links within a bioregion
+    sd_link_bioregion <- tapply(bipartite_df$n_link_bioregion,
+                                bipartite_df$Bioregion,
+                                stats::sd)
+    sd_link_bioregion <-
+      data.frame(Bioregion = names(sd_link_bioregion),
+                 sd_link_bioregion = as.numeric(sd_link_bioregion))
+    bipartite_df <- dplyr::left_join(bipartite_df, sd_link_bioregion,
+                                     by = "Bioregion")
+    
+    # z
+    bipartite_df$z <- (bipartite_df$n_link_bioregion -
+                         bipartite_df$mean_link_bioregion) /
+      bipartite_df$sd_link_bioregion
+    
+    # Remove intermediate columns
+    bipartite_df <-
+      bipartite_df[, c("Node", "Bioregion", "Category", "C", "z")]
   }
   
   ## 2.2. Contribution --------------------------------------------------------
@@ -112,6 +283,13 @@ contribution <- function(cluster_object,
     # Binary site-species matrix
     comat_bin <- comat
     comat_bin[comat_bin > 0] <- 1
+    
+    # If it is a bipartite object, we just consider the sites
+    if(cluster_object$inputs$bipartite == TRUE){
+      cluster_object$clusters <-
+        cluster_object$clusters[
+          which(attributes(cluster_object$clusters)$node_type == "site"), ]
+    }
     
     # Formula
     n <- nrow(comat) # number of sites
@@ -164,14 +342,14 @@ contribution <- function(cluster_object,
   
   if(length(indices) == 1){
     if(indices == "Cz"){
-      return(Cz_df)
+      return(bipartite_df)
     } else if(indices == "contribution"){
       return(contrib_df)    
     }
   } else{
-    if(indices == c("contribution", "Cz") || indices == c("Cz", "contribution")){
+    if("contribution" %in% indices && "Cz" %in% indices){
       return(list(contribution = contribution_df,
-                  cz = Cz_df))
+                  cz = bipartite_df))
     }
   }
 }